Update shared memory management and enhance user utilities
- Modify map_shared_pages function to return uint64 instead of int.
- Add string manipulation functions (strcat, itoa) in log_test.c.
- Improve formatting and readability in ls.c.
- Update VSCode settings for better file associations.

+++ user/user.h
@@ -23,7 +23,7 @@ char *sbrk(int);
 int sleep(int);
 int uptime(void);
 // System calls for shared memory management
-int map_shared_pages(void *src_va, int size, int src_pid, int dest_pid);
+uint64 map_shared_pages(void *src_va, int size, int src_pid, int dest_pid);
 int unmap_shared_pages(void *dst_va, int size);
 int getsz(void);
